Extract SlaveController API endpoints to API namespace
